-Fatta pulizia della cartella LOG; -aggiunto parametro booleano al metodo setFile del Logger: se true sovrascrive se false va in append; -ora battery crea un altro file di log con tutto quello che prima stampava a terminale; -sia battery che TestLauncher utilizzano la versione dei tableau con il Log visto che sono intese come casi di TEST.

+++ src/main/java/ALC_Reasoner/LoggerManager.java
@@ -64,12 +64,14 @@ public class LoggerManager {
      * @param name String The name of the new .log file.
      * @param c    Class&lt;?&gt; The class that call the method.
      */
-    public static void setFile(String name, Class<?> c) {
+    public static void setFile(String name, Class<?> c, boolean overW) {
         File file = new File ("LOG/"+c.getName()+"/"+name+".log");
-        try {
-            Files.deleteIfExists(file.toPath());
-        } catch (IOException e) {
-            e.printStackTrace();
+        if(overW){
+            try {
+                Files.deleteIfExists(file.toPath());
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
         System.setProperty("logfile.name","LOG/"+c.getName()+"/"+name+".log");
         enable = true;
